requirements.txt
--editable git+https://github.com/LonamiWebs/Telethon.git#egg=telethon
vk==2.0.2
pysocks
